# JDK-8260273: DataOutputStream writeChars optimization