--- a/src/java.base/share/classes/java/io/DataOutputStream.java
+++ b/src/java.base/share/classes/java/io/DataOutputStream.java
@@ -300,8 +300,9 @@ public class DataOutputStream extends FilterOutputStream implements DataOutput {
         int len = s.length();
         for (int i = 0 ; i < len ; i++) {
             int v = s.charAt(i);
-            out.write((v >>> 8) & 0xFF);
-            out.write((v >>> 0) & 0xFF);
+            writeBuffer[0] = (byte)(v >>> 8);
+            writeBuffer[1] = (byte)(v >>> 0);
+            out.write(writeBuffer, 0, 2);
         }
         incCount(len * 2);
     }