--- a/test/micro/org/openjdk/bench/java/io/DataOutputStreamTest.java
+++ b/test/micro/org/openjdk/bench/java/io/DataOutputStreamTest.java
@@ -37,7 +37,7 @@ import java.util.concurrent.TimeUnit;
 public class DataOutputStreamTest {
 
     public enum BasicType {CHAR, SHORT, INT, STRING}
-    @Param({"CHAR", "SHORT", "INT", /* "STRING"*/}) BasicType basicType;
+    @Param({"CHAR", "SHORT", "INT", "STRING"}) BasicType basicType;
 
     @Param({"4096"}) int size;
     final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream(size);